requirements.txt
requests
beautifulsoup4
redis-py
hiredis
progressbar2
websockets
aioredis
